# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for user registration, authentication, wallet operations and admin management",
        "title": "mFrelance API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/disputes/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Opens a new dispute for a task (only client or accepted freelancer can open)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "disputes"
                ],
                "summary": "Create dispute",
                "parameters": [
                    {
                        "description": "Resolution payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.ResolveDisputeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dispute created successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request or dispute already exists",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Task not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/disputes/get": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns details and messages for a specific dispute",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "disputes"
                ],
                "summary": "Get dispute details",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Dispute ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dispute details",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid dispute ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Dispute not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/disputes/message": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Allows client, freelancer or assigned admin to send a message in a dispute",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "disputes"
                ],
                "summary": "Send dispute message",
                "parameters": [
                    {
                        "description": "Dispute message",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.SendDisputeMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Message sent",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid JSON or request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Dispute not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to send message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/disputes/my": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all disputes where the user is a client or accepted freelancer",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "disputes"
                ],
                "summary": "Get user disputes",
                "responses": {
                    "200": {
                        "description": "User disputes list",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.ResolveDisputeRequest": {
            "type": "object",
            "properties": {
                "dispute_id": {
                    "type": "integer"
                },
                "resolution": {
                    "description": "\"client_won\" или \"freelancer_won\"",
                    "type": "string"
                }
            }
        },
...
        "handlers.SendDisputeMessageRequest": {
            "type": "object",
            "properties": {
                "dispute_id": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}